requirements.txt
python-telegram-bot==21.1.1
solana==0.34.3
solders==0.21.0
websockets>=11.0,<12.0
python-dotenv==1.0.1
httpx==0.27.0
